pyproject.toml
[project]
name = "vlm-ocr"
version = "0.1.0"
description = "VLM OCR project"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "notebook>=7.4.4",
    "ipykernel>=6.30.1",
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "accelerate>=0.26.0",
    "qwen-vl-utils>=0.0.1",
]

[project.optional-dependencies]
quantization = [
    "bitsandbytes>=0.41.0; platform_system != 'Darwin' or platform_machine != 'arm64'"
]

[tool.uv]
dev-dependencies = []